requirements_prod.txt
aiohappyeyeballs==2.6.1
aiohttp==3.13.3
aiofiles==25.1.0
aiosignal==1.4.0
alembic==1.18.4
annotated-doc==0.0.4
antlr4-python3-runtime==4.9.3
anyio==4.12.1
asteroid-filterbanks==0.4.0
attrs==25.4.0
av==16.1.0
certifi==2026.1.4
cffi==2.0.0
charset-normalizer==3.4.4
click==8.3.1
coloredlogs==15.0.1
colorlog==6.10.1
contourpy==1.3.3
ctranslate2==4.6.0
cycler==0.12.1
docopt==0.6.2
einops==0.8.2
fastapi==0.129.0
faster-whisper==1.1.0
filelock==3.24.2
flatbuffers==25.12.19
fonttools==4.61.1
frozenlist==1.8.0
fsspec==2026.2.0
greenlet==3.3.1
h11==0.16.0
hf-xet==1.2.0
httpcore==1.0.9
httpx==0.28.1
huggingface_hub==0.36.2
humanfriendly==10.0
HyperPyYAML==1.2.3
idna==3.11
Jinja2==3.1.6
joblib==1.5.3
julius==0.2.7
kiwisolver==1.4.9
lightning==2.6.1
lightning-utilities==0.15.2
Mako==1.3.10
markdown-it-py==4.0.0
MarkupSafe==3.0.3
matplotlib==3.10.8
mdurl==0.1.2
mpmath==1.3.0
multidict==6.7.1
networkx==3.6.1
nltk==3.9.2
numpy==1.26.4
omegaconf==2.3.0
onnxruntime==1.23.2
optuna==4.7.0
packaging==26.0
pandas==3.0.0
pillow==12.1.1
primePy==1.3
propcache==0.4.1
protobuf==6.33.5
pyannote.audio==3.3.2
pyannote.core==5.0.0
pyannote.database==5.1.3
pyannote.metrics==3.2.1
pyannote.pipeline==3.0.1
pycparser==3.0
pydantic-settings==2.13.0
pydub==0.25.1
Pygments==2.19.2
pyparsing==3.3.2
python-dateutil==2.9.0.post0
python-dotenv==1.2.1
python-multipart==0.0.22
pytorch-lightning==2.6.1
pytorch-metric-learning==2.9.0
PyYAML==6.0.3
regex==2026.1.15
requests==2.32.5
rich==14.3.2
ruamel.yaml==0.18.17
ruamel.yaml.clib==0.2.15
safetensors==0.7.0
scikit-learn==1.8.0
scipy==1.17.0
semver==3.0.4
sentencepiece==0.2.1
setuptools==82.0.0
shellingham==1.5.4
six==1.17.0
sortedcontainers==2.4.0
soundfile==0.13.1
speechbrain==1.0.3
SQLAlchemy==2.0.46
sympy==1.14.0
tabulate==0.9.0
tensorboardX==2.6.4
threadpoolctl==3.6.0
tokenizers==0.19.1
torch==2.2.2
torch-audiomentations==0.12.0
torch_pitch_shift==1.2.5
torchaudio==2.2.2
torchmetrics==1.8.2
tqdm==4.67.3
transformers==4.40.0
typer==0.23.1
typer-slim==0.23.1
typing_extensions==4.15.0
urllib3==2.6.3
uvicorn==0.41.0
whisperx==3.3.1
yarl==1.22.0
